CR notes
