+++ src/mapper/pkg/labelreporter/pod_reconciler.go
@@ -6,10 +6,8 @@ import (
 	"github.com/otterize/intents-operator/src/shared/serviceidresolver"
 	"github.com/otterize/intents-operator/src/shared/serviceidresolver/serviceidentity"
 	"github.com/otterize/network-mapper/src/mapper/pkg/cloudclient"
-	"github.com/otterize/nilable"
 	"github.com/samber/lo"
 	"github.com/sirupsen/logrus"
-	"golang.org/x/exp/slices"
 	corev1 "k8s.io/api/core/v1"
 	ctrl "sigs.k8s.io/controller-runtime"
 	"sigs.k8s.io/controller-runtime/pkg/client"
@@ -82,68 +80,6 @@ func (r *PodReconciler) Reconcile(ctx context.Context, req ctrl.Request) (ctrl.R
 	return ctrl.Result{}, nil
 }
 
-func (r *PodReconciler) syncOnceAllPods(ctx context.Context) error {
-	allNamespaces := &corev1.NamespaceList{}
-	err := r.List(ctx, allNamespaces)
-	if err != nil {
-		return errors.Wrap(err)
-	}
-	for _, namespace := range allNamespaces.Items {
-		err := r.syncPodsInNamespace(ctx, namespace.Name)
-		if err != nil {
-			return errors.Wrap(err)
-		}
-	}
-
-	return nil
-}
-
-func (r *PodReconciler) syncPodsInNamespace(ctx context.Context, namespace string) error {
-	pods := &corev1.PodList{}
-	err := r.List(ctx, pods, client.InNamespace(namespace))
-	if err != nil {
-		return errors.Wrap(err)
-	}
-
-	serviceIdentityToReportInput := make(map[serviceIdentityKey]cloudclient.ReportServiceMetadataInput)
-
-	for _, pod := range pods.Items {
-		serviceIdentity, err := r.serviceIDResolver.ResolvePodToServiceIdentity(ctx, &pod)
-		if err != nil {
-			return errors.Wrap(err)
-		}
-		if _, ok := serviceIdentityToReportInput[serviceIdentityToCacheKey(serviceIdentity)]; ok {
-			continue
-		}
-		identityInput := serviceIdentityToServiceIdentityInput(serviceIdentity)
-		labelsInput := r.labelsToLabelInput(pod.Labels)
-		input := cloudclient.ReportServiceMetadataInput{
-			Identity: identityInput,
-			Metadata: cloudclient.ServiceMetadataInput{Labels: labelsInput},
-		}
-		serviceIdentityToReportInput[serviceIdentityToCacheKey(serviceIdentity)] = input
-	}
-
-	inputs := lo.Values(serviceIdentityToReportInput)
-	slices.SortFunc(inputs, func(a, b cloudclient.ReportServiceMetadataInput) bool {
-		return a.Identity.Name < b.Identity.Name
-	})
-
-	err = r.cloudClient.ReportWorkloadsLabels(ctx, inputs)
-	if err != nil {
-		return errors.Wrap(err)
-	}
-
-	for key, input := range serviceIdentityToReportInput {
-		labels := lo.SliceToMap(input.Metadata.Labels, func(label cloudclient.LabelInput) (string, string) {
-			return label.Key, label.Value.Item
-		})
-		r.cache.Add(key, checksumLabels(labels))
-	}
-
-	return nil
-}
-
 func (r *PodReconciler) reportWorkloadLabelsWithCache(ctx context.Context, serviceIdentity serviceidentity.ServiceIdentity, labels map[string]string) error {
 	svcIDKey, labelVal := serviceIdentityToCacheKey(serviceIdentity), checksumLabels(labels)
 	cached := r.cache.IsCached(svcIDKey, labelVal)
@@ -163,7 +99,7 @@ func (r *PodReconciler) reportWorkloadLabelsWithCache(ctx context.Context, servi
 func (r *PodReconciler) reportWorkloadLabels(ctx context.Context, serviceIdentity serviceidentity.ServiceIdentity, labels map[string]string) error {
 	serviceIdentityInput := serviceIdentityToServiceIdentityInput(serviceIdentity)
 
-	labelsInput := r.labelsToLabelInput(labels)
+	labelsInput := labelsToLabelInput(labels)
 
 	workloadLabelInput := cloudclient.ReportServiceMetadataInput{
 		Identity: serviceIdentityInput,
@@ -171,20 +107,4 @@ func (r *PodReconciler) reportWorkloadLabels(ctx context.Context, serviceIdentit
 	}
 
 	return errors.Wrap(r.cloudClient.ReportWorkloadsLabels(ctx, []cloudclient.ReportServiceMetadataInput{workloadLabelInput}))
-}
-
-func (r *PodReconciler) labelsToLabelInput(labels map[string]string) []cloudclient.LabelInput {
-	labelsInput := make([]cloudclient.LabelInput, 0)
-	for key, value := range labels {
-		labelsInput = append(labelsInput, cloudclient.LabelInput{Key: key, Value: nilable.From(value)})
-	}
-
-	slices.SortFunc(labelsInput, func(a, b cloudclient.LabelInput) bool {
-		return a.Key < b.Key
-	})
-	return labelsInput
-}
-
-func serviceIdentityToCacheKey(identity serviceidentity.ServiceIdentity) serviceIdentityKey {
-	return serviceIdentityKey(identity.GetNameWithKind())
 }
